Rename featureSlider to FeatureSelector
