+++ src/components/FeatureSelector/AudioFeatureSlider.tsx
