+++ src/components/FeatureSelector/FeatureSliderItem.tsx
@@ -1,0 +1,8 @@
+export interface FeatureSliderItem {
+  name: string
+  min: number
+  max: number
+  currentMin: number
+  currentMax: number
+  units: string
+}
