--- src/types/FeatureSliderItem.tsx
@@ -1,8 +1,0 @@
-export interface FeatureSliderItem {
-    name: string
-    min: number
-    max: number
-    currentMin: number
-    currentMax: number
-    units: string
-  }
